backend/services/item_bank_service/item_bank_service/requirements.txt
fastapi>=0.115.0
sqlmodel>=0.0.21
pydantic-settings>=2.0.0
pyjwt>=2.8.0
openpyxl>=3.1.0
python-multipart>=0.0.9
uvicorn>=0.30.0

# Database drivers (choose one)
# For development/testing:
# sqlite is built into Python

# For production:
# psycopg2-binary>=2.9.9  # PostgreSQL

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
